Fixed console.log errors for RecipeCards and RecipeDetails
